# TicketManagementSystem/backend/TicketManagementSystem.API/openapi.yaml
openapi: 3.1.0
info:
  title: Ticket Management System API
  description: |
    API RESTful para el Sistema de Gestión de Tickets.
    
    ## Autenticación
    Esta API utiliza **JWT Bearer Token** para autenticación. Para acceder a los endpoints protegidos:
    1. Obtener un token mediante `/api/auth/login`
    2. Incluir el token en el header `Authorization: Bearer {token}`
    
    ## Rate Limiting
    - General: 200 requests/minuto
    - Login: 20 requests/minuto
    - Register: 10 requests/minuto
    
    ## Roles
    - **Admin**: Acceso completo al sistema
    - **Agent**: Gestión de tickets asignados
    - **User**: Crear y ver tickets propios
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: dev@ticketmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Servidor de desarrollo
  - url: https://localhost:5001/api
    description: Servidor de desarrollo (HTTPS)
  - url: https://api.ticketmanagement.com/api
    description: Servidor de producción
...
security:
  - BearerAuth: []
paths:
...
  /roles:
    get:
      tags:
        - Catalogs
      summary: Listar roles
      description: Obtener catálogo de roles disponibles
      operationId: getRoles
      responses:
        '200':
          description: Lista de roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleDto'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # ==================== HEALTH ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido mediante el endpoint `/api/auth/login`.
        
        Formato: `Bearer {token}`
        
        Ejemplo: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    RoleDto:
      type: object
      properties:
        id:
          type: integer
          description: ID del rol
          example: 1
        name:
          type: string
          description: Nombre del rol
          example: "Admin"
        description:
          type: string
          description: Descripción del rol
          example: "Administrador del sistema con acceso completo"

    # ==================== ENUMS ====================
...
    ProblemDetails:
      type: object
      description: Formato de error según RFC 7807
      properties:
        type:
          type: string
          description: URI que identifica el tipo de error
          example: "https://tools.ietf.org/html/rfc7231#section-6.5.1"
        title:
          type: string
          description: Título del error
          example: "Bad Request"
        status:
          type: integer
          description: Código HTTP
          example: 400
        detail:
          type: string
          description: Descripción detallada del error
          example: "El campo 'email' es requerido"
        instance:
          type: string
          description: URI de la instancia del error
          example: "/api/auth/login"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Errores de validación por campo
          example:
            email:
              - "El campo 'email' es requerido"
              - "El formato del email es inválido"
            password:
              - "La contraseña debe tener al menos 6 caracteres"
        traceId:
          type: string
          description: ID de traza para debugging
          example: "00-abc123def456-789xyz-00"